Cleaned MenuUI code, added new menu dialogs.
FileBrowserDialog - A FileBrowser that uses MenuUI can also be structured using a list.
ListDialog - Shows a dialog of a list to select one of the items[Will be used a lot in the editor]
SelectListDIalog - Shows a dialog of a toggle list to select multiple ones[will be used for on_executed list in elements etc.]
ScrollablePanel - Changed a few stuff to allow more things.