Final dev commit
Fixed some network issues/crashes related and not related to custom heists.
Added an attempt to fix the coresequence crash
fixed a crash with mod assets module
readded the WeaponModuleNew to avoid crashes when updating from old BeardLib.
Added checks for elements that require a special unit to be loaded to avoid crashes
Fixed a crash with CopyToAsync and added non async one.
Added alpha support for hex color(utils)
Fixed a crash in QuickAnim and made it a little lighter
Added a sort for ListDialog