Fixed stuff in weapon mod module
Reverted the ID thing (decided not to at the end).
Added a unit check that reverts the based on unit if the unit is not loaded.
guess_unit is now on by default in ver="2" and will guess a unit if there isn't a unit defined.
Fixed hidden not working as intended