Fixed host weapon not getting synched on start
Set last weapon string without the checks, it will have to go through them anyway.
Don't delete last weapon string if the unit is invalid, it may be still not ready for use.
Removed unneeded fallback in unpack_beardlib_weapon_string (this gets checked anyway and if it doesn't exist no reason to try doing anything with that data)