Fixed #371
Note: If you see weapons in third person mode showing up weirdly this is NOT a BeardLib issue. If you are a mod creator, you'll need to manually fix your mod IF one of your cc_thq materials is not setup correctly. Most likely fixable by adding DEPTH_SCALING to it.