Fix Hardcoded element Width in Achievements List
Makes the width of the achievement Button in the achievement List adapt to the panel width, instead of having a hardcoded value. The GridMenu's wrapping behaviour otherwise would put all the icons of the 2nd row and following rows after the previous row. Now, the elements always fill the whole row, thus the wrapping works correctly.