Add Github provider
a version file is required for commit updates otherwise it will always show as needing updates

release updates use the release's tag, so that should match the version.
a zip of the folder should be the first file in the release.

params:
id = string, owner/name of the repo. ex. Makowo/Beardlib-Editor
branch = string, targeted branch. unused if release is true
release = bool, if set to true, will try to get the latest release instead of commit.