[OptionModule] New multichoice parameter: use_value
An alternative to save_value that works better, it will use the value everywhere, including in .value, this is more consistent with what's saved in the options JSON and does not break older mods. If you want to use save_value, use this instead. save_value had a few weird issues.