Fix custom music sometimes not playing
This was caused by the MusicManager using its original check_music_switch function when the initially chosen music is not custom music.
However, doing it like this allows the vanilla function to pick a random track again which may result in it picking a custom track and BeardLib not attempting to play it.
Simply post hooking the function instead to attempt to play whatever the vanilla function picked is enough, as a music switch to a custom track in the vanilla function won't play anything anyways and the sound source can also be stopped.